requirements.txt
dnspython>=2.4.2
requests>=2.31.0
